+++ src/attributes.ts
@@ -1,4 +1,5 @@
 import { MarkdownString, Uri } from "vscode";
+import * as fs from "fs";
 
 type ToolType = "axe" | "hoe" | "pickaxe" | "shovel" | "sword" | "shears" | "trident" | "bow" | "crossbow" | "shield" | "armor";
 
@@ -26,7 +27,8 @@ export class ItemAttribute {
     private getImagePath(): string {
         let [base, loc] = this.id.split(":");
         loc = loc.replace("/", "_");
-        return `./kubejs/probe/cache/rich/${base}/${loc}.png`;
+        let path = `./kubejs/probe/cache/rich/${base}/${loc}.png`;
+        return fs.existsSync(path) ? path : `./kubejs/probe/cache/rich/item/${base}/${loc}.png`;
     }
 
     private getDescription(): string {
@@ -91,17 +93,22 @@ export class TagAttribute {
     id!: `${string}:${string}`;
     items!: string[];
 
-    private getImagePath(itemId: string): string {
-        let [base, loc] = itemId.split(":");
+    private getImagePath(id: string): string {
+        let [base, loc] = id.split(":");
         loc = loc.replace("/", "_");
-        return `./kubejs/probe/cache/rich/${base}/${loc}.png`;
+        let path = `./kubejs/probe/cache/rich/${base}/${loc}.png`;
+        return fs.existsSync(path) ? path : `./kubejs/probe/cache/rich/item/${base}/${loc}.png`;
     }
 
     private getImageMarkdowns(): string {
         let itemMds = this.items.map((item) => `![${item}](${this.getImagePath(item)})`);
-        // clamp to 5 items, if clamped, add a "..." at the end
-        if (itemMds.length > 5) {
-            itemMds = itemMds.slice(0, 5);
+        // start a new line for every 6 items, max at 12 in total
+        for (let i = 6; i < itemMds.length; i += 6) {
+            itemMds[i] = "\n\n" + itemMds[i];
+        }
+        // clamp the length to 12, add ... if there are more
+        if (itemMds.length > 12) {
+            itemMds.length = 12;
             itemMds.push("**...**");
         }
         return itemMds.join(" ");
@@ -111,7 +118,93 @@ export class TagAttribute {
         const md = new MarkdownString("### " + this.id);
         md.baseUri = baseUri;
         md.appendMarkdown(`\n\n${this.getImageMarkdowns()}`);
-        md.appendMarkdown(`\n\n**Count**: ${this.items.length}`)
+        md.appendMarkdown(`\n\n**Count**: ${this.items.length}`);
+        return md;
+    }
+}
+
+export class FluidAttribute {
+    id!: `${string}:${string}`;
+    localized?: string;
+    hasBucket?: boolean;
+    hasBlock?: boolean;
+
+    bucketItem?: `${string}:${string}`;
+
+    private getImagePath(): string {
+        let [base, loc] = this.id.split(":");
+        loc = loc.replace("/", "_");
+        return `./kubejs/probe/cache/rich/fluid/${base}/${loc}.png`;
+    }
+
+    private getBucketImagePath(): string {
+        let [base, loc] = this.bucketItem!.split(":");
+        loc = loc.replace("/", "_");
+        return `./kubejs/probe/cache/rich/item/${base}/${loc}.png`;
+    }
+
+    public getMarkdown(baseUri: Uri): MarkdownString {
+        const md = new MarkdownString("### " + this.localized ?? this.id);
+        md.baseUri = baseUri;
+
+        md.appendMarkdown(`\n\n![Image](${this.getImagePath()})`);
+
+        if (this.localized) {
+            md.appendMarkdown(`\n\n**ID**: ${this.id}`);
+        }
+
+        if (this.hasBlock) {
+            md.appendMarkdown(`\n\nThis fluid has a **block**.`);
+        }
+
+        if (this.hasBucket) {
+            md.appendMarkdown(`\n\nThis fluid has a **bucket item**:`);
+            md.appendMarkdown(`\n\n![Image](${this.getBucketImagePath()})`);
+        }
+
+        return md;
+    }
+}
+
+export class LanguageAttribute {
+    key!: string;
+    selected!: string;
+    languages!: { [key: string]: string };
+
+    public getMarkdown(): MarkdownString {
+        const md = new MarkdownString("", true);
+
+        let missingTranslations = [];
+
+        // append a table with `lang` and `translation`
+        md.appendMarkdown(`| Language | Translation |\n`);
+        md.appendMarkdown(`| -------- | ----------- |\n`);
+        // append the selected language
+        md.appendMarkdown(`| **${this.selected}** | **${this.languages[this.selected]}** |\n`);
+
+        if (this.languages[this.selected] === this.languages['en_us']) {
+            missingTranslations.push(this.selected);
+        }
+
+        // append the rest of the languages
+        for (const lang in this.languages) {
+            if (lang !== this.selected && lang !== "en_us") {
+                if (this.languages[lang] !== this.languages['en_us']) {
+                    md.appendMarkdown(`| ${lang} | ${this.languages[lang]} |\n`);
+                } else {
+                    missingTranslations.push(lang);
+                }
+            }
+        }
+        // append the en_us as fallback
+        if (this.selected !== "en_us") {
+            md.appendMarkdown(`| en_us | ${this.languages["en_us"]} |\n`);
+        }
+
+        if (missingTranslations.length > 0 && !this.selected.startsWith("en")) {
+            md.appendMarkdown(`\n\n**Missing Translations: ${missingTranslations.join(", ")}**`);
+        }
+
         return md;
     }
 }